+++ modules/tinymce/src/core/main/ts/api/Options.ts
@@ -381,6 +381,11 @@ const register = (editor: Editor): void => {
     default: false
   });
 
+  registerOption('format_noneditable_selector', {
+    processor: 'string',
+    default: ''
+  });
+
   registerOption('preview_styles', {
     processor: (value) => {
       const valid = value === false || Type.isString(value);
@@ -837,6 +842,7 @@ const isInlineBoundariesEnabled = option('inline_boundaries');
 const getFormats = option('formats');
 const getPreviewStyles = option('preview_styles');
 const canFormatEmptyLines = option('format_empty_lines');
+const getFormatNoneditableSelector = option('format_noneditable_selector');
 const getCustomUiSelector = option('custom_ui_selector');
 const isInline = option('inline');
 const hasHiddenInput = option('hidden_input');
@@ -943,6 +949,7 @@ export {
   getFormats,
   getPreviewStyles,
   canFormatEmptyLines,
+  getFormatNoneditableSelector,
   getCustomUiSelector,
   getThemeUrl,
   getModelUrl,