+++ modules/tinymce/src/core/main/ts/fmt/MergeUtils.ts
@@ -12,7 +12,7 @@ import * as FormatUtils from './FormatUtils';
 
 const each = Tools.each;
 
-const isElementNode = (node: Node): node is Element =>
+const isElementNode = (node: Node): node is HTMLElement =>
   NodeType.isElement(node) && !Bookmarks.isBookmarkNode(node) && !isCaretNode(node) && !NodeType.isBogus(node);
 
 const findElementSibling = (node: Node, siblingName: 'nextSibling' | 'previousSibling') => {
@@ -31,9 +31,11 @@ const findElementSibling = (node: Node, siblingName: 'nextSibling' | 'previousSi
 
 const mergeSiblingsNodes = (editor: Editor, prev: Node | undefined, next: Node | undefined) => {
   const elementUtils = ElementUtils(editor);
+  const isPrevEditable = NodeType.isElement(prev) && FormatUtils.isEditable(prev);
+  const isNextEditable = NodeType.isElement(next) && FormatUtils.isEditable(next);
 
   // Check if next/prev exists and that they are elements
-  if (prev && next) {
+  if (isPrevEditable && isNextEditable) {
     // If previous sibling is empty then jump over it
     const prevSibling = findElementSibling(prev, 'previousSibling');
     const nextSibling = findElementSibling(next, 'nextSibling');
@@ -62,6 +64,7 @@ const mergeSiblingsNodes = (editor: Editor, prev: Node | undefined, next: Node |
 
 const mergeSiblings = (editor: Editor, format: ApplyFormat, vars: FormatVars | undefined, node: Node): void => {
   // Merge next and previous siblings if they are similar <b>text</b><b>text</b> becomes <b>texttext</b>
+  // Note: mergeSiblingNodes attempts to not merge sibilings if they are noneditable
   if (node && format.merge_siblings !== false) {
     // Previous sibling
     const newNode = mergeSiblingsNodes(editor, FormatUtils.getNonWhiteSpaceSibling(node), node) ?? node;
@@ -73,17 +76,17 @@ const mergeSiblings = (editor: Editor, format: ApplyFormat, vars: FormatVars | u
 const clearChildStyles = (dom: DOMUtils, format: ApplyFormat, node: Node): void => {
   if (format.clear_child_styles) {
     const selector = format.links ? '*:not(a)' : '*';
-    each(dom.select(selector, node), (node) => {
-      if (isElementNode(node)) {
-        each(format.styles, (value, name: string) => {
-          dom.setStyle(node, name, '');
+    each(dom.select(selector, node), (childNode) => {
+      if (isElementNode(childNode) && FormatUtils.isEditable(childNode)) {
+        each(format.styles, (_value, name: string) => {
+          dom.setStyle(childNode, name, '');
         });
       }
     });
   }
 };
 
-const processChildElements = (node: Node, filter: (element: Element) => boolean, process: (element: Element) => void): void => {
+const processChildElements = (node: Node, filter: (element: HTMLElement) => boolean, process: (element: HTMLElement) => void): void => {
   each(node.childNodes, (node) => {
     if (isElementNode(node)) {
       if (filter(node)) {